core stuff
.
added a clearlist for nextwave, and also destroy happens when next wave is spawned instead of destroying each object individually, saves some GC but tbh its much better to not confuse our lists by destroying actively, thanks to states i can keep them in dead state in my list and still filter through.
So, i clear all list of enemies on nextwave and also destroy all GO in the list, probably dead.

i was mistakenly playing with my list from bullet script as well. it was supposed to be a test trigger but i forgot to comment it out